+++ cmd/project.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/pclubiitk/fep-backend/middleware"
+	//"github.com/pclubiitk/fep-backend/middleware"
 	"github.com/pclubiitk/fep-backend/project"
 	"github.com/spf13/viper"
 )
@@ -12,12 +12,13 @@ import (
 func adminProjectServer() *http.Server {
 	PORT := viper.GetString("PORT.PROJECT")
 	engine := gin.New()
-	engine.Use(middleware.CORS())
-	engine.Use(middleware.Authenticator())
-	engine.Use(middleware.EnsureAdmin())
-	engine.Use(gin.Recovery())
-	engine.Use(gin.Logger())
+	// engine.Use(middleware.CORS())
+	// engine.Use(middleware.Authenticator())
+	// engine.Use(middleware.EnsureAdmin())
+	// engine.Use(gin.Recovery())
+	// engine.Use(gin.Logger())
 	project.AdminRouter(engine)
+	project.StudentRouter(engine)
 
 	server := &http.Server{
 		Addr:         ":" + PORT,
