+++ project/router.go
@@ -19,3 +19,9 @@ func AdminRouter(r *gin.Engine) {
 
 	}
 }
+func StudentRouter(r *gin.Engine) {
+    student := r.Group("/api/student/application")
+    {
+        student.POST("", addApplicationHandler)  // Route for students to add an application
+    }
+}
